pyproject.toml
[project]
name = "pyxel-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.8.3",
    "lxml>=6.0.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyxel>=2.5.4",
    "requests>=2.10.0",
    "tqdm>=4.67.1",
    "typing>=3.10.0.0",
    "yfinance>=0.2.65",
]
